piece crud

+++ database/migrations/2024_05_21_191951_create_order_details.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('description');
-            $table->unsignedBigInteger('price_unit');
+            $table->decimal('price_unit', 10, 2);
 
             $table->unsignedBigInteger('order_id');
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
